Move native libraries build to cmake (#2009)
