+++ horovod/common/ops/cuda/CMakeLists.txt
@@ -1,23 +1,16 @@
-cmake_minimum_required(VERSION 2.8 FATAL_ERROR)
+set(CUDA_HOST_COMPILER ${CMAKE_CXX_COMPILER})
 
-project(horovod_cuda_kernels)
-
-find_package(CUDA)
-
-# Remove C++ std flags (--std=c++11 included in CUDA_NVCC_FLAGS)
-string(REGEX REPLACE  "-std=[^ ]+" "" CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}")
-
-# Set ar
-if(DEFINED ENV{AR})
-  set(CMAKE_AR $ENV{AR})
-endif()
+set(ENV{PYTHONPATH} "${PROJECT_SOURCE_DIR}/cmake:$ENV{PYTHONPATH}")
+execute_process(COMMAND ${PY_EXE} -c "import build_utils; print(' '.join(build_utils.get_nvcc_flags()))"
+                OUTPUT_VARIABLE HVD_NVCC_COMPILE_FLAGS OUTPUT_STRIP_TRAILING_WHITESPACE)
 
 MESSAGE(STATUS "HVD_NVCC_COMPILE_FLAGS = ${HVD_NVCC_COMPILE_FLAGS}")
 
-list(APPEND CUDA_NVCC_FLAGS ${HVD_NVCC_COMPILE_FLAGS})
+list(APPEND CUDA_NVCC_FLAGS "${HVD_NVCC_COMPILE_FLAGS}")
 
-cuda_add_library(horovod_cuda_kernels
-    cuda_kernels.cu
-)
+cuda_add_library(horovod_cuda_kernels cuda_kernels.cu)
+target_compile_definitions(horovod_cuda_kernels PRIVATE _GLIBCXX_USE_CXX11_ABI=1)
 
-set(CMAKE_VERBOSE_MAKEFILE ON)
+# if we need compatible c++ abi, build a compatible version
+cuda_add_library(compatible_horovod_cuda_kernels cuda_kernels.cu)
+target_compile_definitions(compatible_horovod_cuda_kernels PRIVATE _GLIBCXX_USE_CXX11_ABI=0)
